+++ packages/jest-core/src/__tests__/SearchSource.test.ts
@@ -10,7 +10,7 @@ import * as path from 'path';
 import type {Config} from '@jest/types';
 import {normalize} from 'jest-config';
 import {Test} from 'jest-runner';
-import Runtime = require('jest-runtime');
+import Runtime from 'jest-runtime';
 import SearchSource, {SearchResult} from '../SearchSource';
 
 jest.setTimeout(15000);