+++ packages/jest-core/src/__tests__/watchFileChanges.test.ts
@@ -13,7 +13,7 @@ import rimraf = require('rimraf');
 import type {AggregatedResult} from '@jest/test-result';
 import {normalize} from 'jest-config';
 import HasteMap = require('jest-haste-map');
-import Runtime = require('jest-runtime');
+import Runtime from 'jest-runtime';
 import {JestHook} from 'jest-watcher';
 
 describe('Watch mode flows with changed files', () => {